Refactor Flask app into factory and routes module
- Add create_app() in backend package initialization
- Keep backend.app as a lightweight compatible entry point
- Move routes, handlers, helpers, and R2 functions into backend.routes
- Update tests to mock upload/delete helpers from the new routes module
